test: clean up PolicyEvaluatorTest
