+++ core/policy/policy-engine/src/test/java/org/eclipse/dataspaceconnector/policy/engine/model/PolicyTestFunctions.java
@@ -9,14 +9,26 @@
  *
  *  Contributors:
  *       Microsoft Corporation - initial API and implementation
+ *       Fraunhofer Institute for Software and Systems Engineering - resource definition functions
  *
  */
 
 package org.eclipse.dataspaceconnector.policy.engine.model;
 
+import java.util.stream.Stream;
+
+import com.fasterxml.jackson.annotation.JsonCreator;
+import com.fasterxml.jackson.databind.annotation.JsonPOJOBuilder;
 import org.eclipse.dataspaceconnector.policy.model.AtomicConstraint;
+import org.eclipse.dataspaceconnector.policy.model.Duty;
 import org.eclipse.dataspaceconnector.policy.model.LiteralExpression;
 import org.eclipse.dataspaceconnector.policy.model.Operator;
+import org.eclipse.dataspaceconnector.policy.model.Permission;
+import org.eclipse.dataspaceconnector.policy.model.Prohibition;
+import org.eclipse.dataspaceconnector.spi.types.domain.transfer.ResourceDefinition;
+import org.junit.jupiter.api.extension.ExtensionContext;
+import org.junit.jupiter.params.provider.Arguments;
+import org.junit.jupiter.params.provider.ArgumentsProvider;
 
 /**
  * Functions used for testing.
@@ -30,5 +42,45 @@ public class PolicyTestFunctions {
         LiteralExpression left = new LiteralExpression(value1);
         LiteralExpression right = new LiteralExpression(value2);
         return AtomicConstraint.Builder.newInstance().leftExpression(left).operator(Operator.EQ).rightExpression(right).build();
+    }
+    
+    public static class TestDefinition extends ResourceDefinition {
+        String key = "someValue";
+    
+        public String getKey() {
+            return this.key;
+        }
+    
+        public void setKey(String key) {
+            this.key = key;
+        }
+    
+        @JsonPOJOBuilder(withPrefix = "")
+        public static class Builder extends ResourceDefinition.Builder<TestDefinition, TestDefinition.Builder> {
+            private Builder() {
+                super(new TestDefinition());
+            }
+        
+            @JsonCreator
+            public static TestDefinition.Builder newInstance() {
+                return new TestDefinition.Builder();
+            }
+        
+            public TestDefinition.Builder key(String key) {
+                resourceDefinition.key = key;
+                return this;
+            }
+        }
+    }
+    
+    public static class ResourceDefinitionFunctionsArguments implements ArgumentsProvider {
+        @Override
+        public Stream<? extends Arguments> provideArguments(ExtensionContext context) {
+            return Stream.of(
+                    Arguments.arguments(Permission.class),
+                    Arguments.arguments(Prohibition.class),
+                    Arguments.arguments(Duty.class)
+            );
+        }
     }
 }
